requirements.txt
Django==1.10.6
psycopg2==2.7
django-webpack-loader==0.4.1
django-oauth-toolkit==0.12.0
markdown==2.6.8
djangorestframework==3.6.2
django-axes==2.3.2
django_filter==1.0.1
Pillow==4.0.0
pytz==2016.10
django-otp==0.4.1.1
django-formtools==2.0
django-two-factor-auth==1.6.2
geoip2==2.5.0
django-user-sessions==1.5.3
